backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1
hiredis==2.2.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
scipy>=1.11.0
faiss-cpu>=1.7.4

# Machine Learning
joblib>=1.3.0

# Math & Scientific Computing
sympy==1.12
scipy==1.11.4

# NLP & ML
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Vector Database
qdrant-client>=1.7.0
hnswlib>=0.8.0

# Search Engine
elasticsearch==8.11.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Configuration & Environment
pydantic[email]==2.5.2
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8